pyproject.toml
[project]
name = "claudius"
version = "0.1.0"
description = "An autonomous AI companion built on Claude Code — always-on daemon with messaging, skills, memory, and multi-agent delegation"
authors = [
    {name = "Andreas von Stegen", email = "ajoechl@pm.me"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10",
    "aiosqlite>=0.20",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "watchdog>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
